requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
pydantic==2.11.7
pydantic-settings==2.10.1
google-api-python-client==2.177.0
google-auth==2.40.3
supabase==2.18.1